Move the dialog code to its own file
