+++ Projects/my_notes/lib/utilities/show_erro_dialog.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+Future<void> showErrorDialog(BuildContext context, String text) {
+  return showDialog(
+      context: context,
+      builder: ((context) {
+        return AlertDialog(
+          title: const Text("An error occured"),
+          content: Text(text),
+          actions: [
+            TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                },
+                child: Text("Okay!"))
+          ],
+        );
+      }));
+}
