+++ Projects/my_notes/lib/view/login_view.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:my_notes/code_repeat.dart';
 
 import 'package:my_notes/constants/routes.dart';
+import 'package:my_notes/utilities/show_erro_dialog.dart';
 
 class LoginView extends StatefulWidget {
   const LoginView({Key? key}) : super(key: key);
@@ -88,22 +89,4 @@ class _LoginViewState extends State<LoginView> {
       ),
     );
   }
-}
-
-Future<void> showErrorDialog(BuildContext context, String text) {
-  return showDialog(
-      context: context,
-      builder: ((context) {
-        return AlertDialog(
-          title: const Text("An error occured"),
-          content: Text(text),
-          actions: [
-            TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-                child: Text("Okay!"))
-          ],
-        );
-      }));
 }
